Work on base types

+++ Universe/src/engine/lib/dt/base/uint64.h
@@ -8,7 +8,7 @@
 
 namespace engine::lib::dt::base
 {
-    using uint32 = std::uint32_t;
+    using uint64 = std::uint64_t;
 
     /*template<>
     struct hash<uint32>
